Changes to the master list.
I changed the master list from a List to a Dictionary. The use pattern fit Dictionary quite closely.

I changed the way we connect other objects to the masterList. I added a static method that finds the Master object and gets the masterList from there. I added code to get the master list in the Start() method of several components.